Adding ipuNumber validation
Summary: - Checks that the ipuNumber is valid when prepareing the devicex

Test Plan: test.sh all - passes

Reviewers: davidn, evan

Reviewed By: evan

Maniphest Tasks: T6122

Differential Revision: https://phabricator.sourcevertex.net/D8450

+++ willow/src/popx/op/ipucopyx.cpp
@@ -18,7 +18,7 @@ void IpuCopyOpx::grow(poplar::program::Sequence &prog) const {
 
   insert(
       outId(0),
-      poputil::copyToIpu(masterGraph(), get(inId(0)), prog, op.getDestIpu()));
+      poputil::copyToIpu(masterGraph(), get(inId(0)), prog, static_cast<int>(op.getDestIpu())));
 }
 
 namespace {
